+++ test/bigtest/tests/declare-lost-test.js
@@ -186,6 +186,58 @@ describe('Declare Lost', () => {
     });
   });
 
+  describe('Fine incurred after marking item as lost', () => {
+    let refreshCounter = 0;
+
+    beforeEach(async function () {
+      const loan = this.server.create('loan', {
+        status: { name: 'Open' },
+        loanPolicyId: 'test',
+        action: 'claimedReturned',
+        actionComment: 'Claim returned confirmation',
+        item: { status: { name: 'Claimed returned' } },
+        itemStatus: 'Claimed returned',
+        claimedReturnedDate: new Date().toISOString(),
+      });
+
+      this.server.create('user', { id: loan.userId });
+      this.server.create('loanaction', {
+        loan: {
+          ...loan.attrs,
+        },
+      });
+
+      // This is ugly but it allows for moving forward without mocking up
+      // too many other backend things.
+      this.server.get('/accounts', () => {
+        const accounts = (refreshCounter > 3) ? [{ id: 1, amount: 250 }] : [];
+        refreshCounter++;
+        return { accounts };
+      });
+
+      this.visit(`/users/${loan.userId}/loans/view/${loan.id}`);
+
+      await LoanActionsHistory.whenLoaded();
+    });
+
+    it('should show empty fine incurred amount', () => {
+      expect(LoanActionsHistory.feeFines.text).to.equal('-');
+    });
+
+    describe('Update fine incurred', () => {
+      beforeEach(async function () {
+        await LoanActionsHistory.declareLostButton.click();
+        await OpenLoansInteractor.declareLostDialog.additionalInfoTextArea.focus();
+        await OpenLoansInteractor.declareLostDialog.additionalInfoTextArea.fill('item lost');
+        await OpenLoansInteractor.declareLostDialog.confirmButton.click();
+      });
+    });
+
+    it('should update fine incurred amount', () => {
+      expect(LoanActionsHistory.feeFines.text).to.equal('250.00');
+    });
+  });
+
   describe('Visiting loan details page with declared lost item', () => {
     beforeEach(async function () {
       const loan = this.server.create('loan', {